pyproject.toml
[tool.poetry]
name = "rilato"
version = "1.0.2"
description = "News reader for GNOME"
authors = ["Gabriele Musco <gabmus@disroot.org>"]
license = "GPL3"

[tool.poetry.dependencies]
python = "^3.10"
Pillow = "^10.4.0"
python-magic = "^0.4.27"
requests = "^2.28.2"
python-dateutil = "^2.8.2"
Pygments = "^2.14.0"
pytz = "^2022.7.1"
readability-lxml = "^0.8.1"
humanize = "^4.6.0"
beautifulsoup4 = "^4.11.2"
lxml = "^5.4.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-cov = "^6.2.1"
ruff = "^0.12.0"
pre-commit = "^3.6.2"

[tool.poetry.scripts]
rilato-build = "tasks:main"
rilato = "tasks:run"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [
  "."
]

[tool.pylsp-mypy]
enabled = true
